Fix function declaration statement body
By having the body be a block statement, there were some additional
unwanted implicit semantics: we both open up a scope for a block
statement and for a function body, so in this case we were having an
additional scope being added!

Switching it up to a list of statements makes more sense.